rework magnum install script to deploy magnum from source code

+++ files/magnum/setup_magnum.sh
@@ -1,4 +1,10 @@
 #!/bin/bash
+if [ -f /var/lib/magnum/magnum-installed ]
+then
+	echo "Magnum is installed - ignore install script!"
+	exit 0
+fi
+
 apt-get -y install python-dev libssl-dev libxml2-dev python-setuptools \
                    libmysqlclient-dev libxslt-dev libpq-dev git \
                    libffi-dev gettext build-essential gcc
@@ -46,12 +52,21 @@ cp doc/examples/etc/systemd/system/magnum-api.service /etc/systemd/system/magnum
 cp doc/examples/etc/systemd/system/magnum-conductor.service /etc/systemd/system/magnum-conductor.service
 cp doc/examples/etc/logrotate.d/magnum.logrotate /etc/logrotate.d/magnum
 
+#symlink magnum-api & magnum-conductor ... to /usr/bin
+ln -s /var/lib/magnum/env/bin/magnum-api /usr/bin/magnum-api
+ln -s /var/lib/magnum/env/bin/magnum-conductor /usr/bin/magnum-conductor
+ln -s /var/lib/magnum/env/bin/magnum-db-manage /usr/bin/magnum-db-manage
+ln -s /var/lib/magnum/env/bin/magnum-template-manage /usr/bin/magnum-template-manage
+
 systemctl enable magnum-api
 systemctl enable magnum-conductor
 
 # Magnum Client
 curl "https://bootstrap.pypa.io/get-pip.py" -o "/tmp/get-pip.py"
 python /tmp/get-pip.py
+
+#TODO - install magnum client in virtual environment
+#TODO - symlink magnum to /usr/bin
 
 cd /opt/
 git clone https://git.openstack.org/openstack/python-magnumclient.git magnumclient
@@ -62,3 +77,5 @@ python setup.py install
 
 ## Create Zertification Directory
 mkdir -p /var/lib/magnum/certificates/
+
+touch /var/lib/magnum/magnum-installed
